Move `unique_selected_lines` to `ui.py`
